Move resource out of parser
